📝 Feat: "추가 카카오 회원가입"

See merge request s06-bigdata-dist-sub2/S06P22A406!24

+++ backend/src/main/java/com/newsbig/sinmunmul/entity/News.java
@@ -33,7 +33,7 @@ public class News implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "news_seq")
-	private int newsSeq;
+	private long newsSeq;
 	
 	@Column(name = "news_title")
 	private String newsTitle;
